requirements.txt
# Spark & Delta
pyspark==3.5.0
delta-spark==3.1.0

# Kafka producer
kafka-python==2.0.2

# Cosmos DB SDK
azure-cosmos==4.7.0

# Testing
pytest==8.2.0
pytest-cov==5.0.0
pytest-html==4.1.1
chispa==0.9.4
httpx==0.27.0

# Linting
ruff==0.4.4
black==24.4.2
isort==5.13.2

# Utilities
python-dotenv==1.0.1
urllib3==2.2.1
